Refactoring the Repositories

+++ src/main/java/com/baha/TrainingPlatformEE/Repositories/SkierRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SkierRepository extends JpaRepository<Skier,Integer> {
+public interface SkierRepository extends JpaRepository<Skier,Long> {
     //extends JpaRepository<Integer,Skier>
     //I  think i found the problem it is implementation problem !
     //next time , get to the method to check for the parameters !
